+++ doorstop/core/item.py
@@ -4,22 +4,22 @@
 
 import functools
 import os
-import re
 from typing import Any, List
 
 from doorstop import common, settings
-from doorstop.common import DoorstopError, DoorstopInfo, DoorstopWarning
+from doorstop.common import DoorstopError
 from doorstop.core import editor
 from doorstop.core.base import (
     BaseFileObject,
-    BaseValidatable,
     add_item,
     auto_load,
     auto_save,
     delete_item,
     edit_item,
 )
+from doorstop.core.reference_finder import ReferenceFinder
 from doorstop.core.types import UID, Level, Prefix, Stamp, Text, to_bool
+from doorstop.core.yaml_validator import YamlValidator
 
 log = common.logger(__name__)
 
@@ -50,6 +50,32 @@ def _convert_to_yaml(indent, prefix, value):
     return value
 
 
+def _convert_to_str(value, result):
+    """Convert value to a string serialization.
+
+    This function is independent of the YAML format and may be used for data
+    which should be independent of the actual item storage format.  It depends
+    only on the Python sorting function, type information, and string
+    representation.
+
+    :param value: the value to convert
+    :param result: the current result of the string serialization
+
+    :return: the updated result of the string serialization
+    """
+    if isinstance(value, list):
+        result += "\\L"
+        for v in value:
+            result = _convert_to_str(v, result)
+        return result
+    if isinstance(value, dict):
+        result += "\\D"
+        for k in sorted(value.keys()):
+            result = _convert_to_str(value[k], result)
+        return result
+    return result + "\\T" + str(type(value)) + "\\V" + str(value).replace("\\", "\\\\")
+
+
 def requires_tree(func):
     """Require a tree reference."""
 
@@ -64,7 +90,7 @@ def requires_tree(func):
     return wrapped
 
 
-class Item(BaseValidatable, BaseFileObject):  # pylint: disable=R0902
+class Item(BaseFileObject):  # pylint: disable=R0902
     """Represents an item file with linkable text."""
 
     EXTENSIONS = '.yml', '.yaml'
@@ -103,10 +129,12 @@ class Item(BaseValidatable, BaseFileObject):  # pylint: disable=R0902
             raise DoorstopError(msg)
         # Initialize the item
         self.path = path
-        self.root = root
+        self.root: str = root
         self.document = document
         self.tree = kwargs.get('tree')
         self.auto = kwargs.get('auto', Item.auto)
+        self.reference_finder = ReferenceFinder()
+        self.yaml_validator = YamlValidator()
         # Set default values
         self._data['level'] = Item.DEFAULT_LEVEL
         self._data['active'] = Item.DEFAULT_ACTIVE
@@ -115,6 +143,7 @@ class Item(BaseValidatable, BaseFileObject):  # pylint: disable=R0902
         self._data['reviewed'] = Item.DEFAULT_REVIEWED
         self._data['text'] = Item.DEFAULT_TEXT
         self._data['ref'] = Item.DEFAULT_REF
+        self._data['references'] = None
         self._data['links'] = set()
         if settings.ENABLE_HEADERS:
             self._data['header'] = Item.DEFAULT_HEADER
@@ -173,6 +202,7 @@ class Item(BaseValidatable, BaseFileObject):  # pylint: disable=R0902
 
     def _set_attributes(self, attributes):
         """Set the item's attributes."""
+        self.yaml_validator.validate_item_yaml(attributes)
         for key, value in attributes.items():
             if key == 'level':
                 value = Level(value)
@@ -188,6 +218,20 @@ class Item(BaseValidatable, BaseFileObject):  # pylint: disable=R0902
                 value = Text(value)
             elif key == 'ref':
                 value = value.strip()
+            elif key == 'references':
+                stripped_value = []
+                for ref_dict in value:
+                    ref_type = ref_dict['type']
+                    ref_path = ref_dict['path']
+
+                    stripped_ref_dict = {"type": ref_type, "path": ref_path.strip()}
+                    if 'keyword' in ref_dict:
+                        ref_keyword = ref_dict['keyword']
+                        stripped_ref_dict['keyword'] = ref_keyword
+
+                    stripped_value.append(stripped_ref_dict)
+
+                value = stripped_value
             elif key == 'links':
                 value = set(UID(part) for part in value)
             elif key == 'header':
@@ -240,6 +284,19 @@ class Item(BaseValidatable, BaseFileObject):  # pylint: disable=R0902
                     value = ''
             elif key == 'ref':
                 value = value.strip()
+            elif key == 'references':
+                if value is None:
+                    continue
+                stripped_value = []
+                for el in value:
+                    ref_dict = {"path": el["path"].strip(), "type": "file"}
+
+                    if 'keyword' in el:
+                        ref_dict['keyword'] = el['keyword']
+
+                    stripped_value.append(ref_dict)
+
+                value = stripped_value
             elif key == 'links':
                 value = [{str(i): i.stamp.yaml} for i in sorted(value)]
             elif key == 'reviewed':
@@ -260,16 +317,6 @@ class Item(BaseValidatable, BaseFileObject):  # pylint: disable=R0902
         """Get the item's UID."""
         filename = os.path.basename(self.path)
         return UID(os.path.splitext(filename)[0])
-
-    @property
-    def prefix(self):
-        """Get the item UID's prefix."""
-        return self.uid.prefix
-
-    @property
-    def number(self):
-        """Get the item UID's number."""
-        return self.uid.number
 
     @property  # type: ignore
     @auto_load
@@ -440,6 +487,20 @@ class Item(BaseValidatable, BaseFileObject):  # pylint: disable=R0902
 
     @property  # type: ignore
     @auto_load
+    def references(self):
+        """Get the item's external file references."""
+        return self._data['references']
+
+    @references.setter  # type: ignore
+    @auto_save
+    def references(self, value):
+        """Set the item's external file references."""
+        if value is not None:
+            assert isinstance(value, list)
+        self._data['references'] = value
+
+    @property  # type: ignore
+    @auto_load
     def links(self):
         """Get a list of the item UIDs this item links to."""
         return sorted(self._data['links'])
@@ -499,7 +560,6 @@ class Item(BaseValidatable, BaseFileObject):  # pylint: disable=R0902
         """Set the item's attributes and save them."""
         self._set_attributes(attributes)
 
-    @auto_save
     def edit(self, tool=None, edit_all=True):
         """Open the item for editing.
 
@@ -513,7 +573,9 @@ class Item(BaseValidatable, BaseFileObject):  # pylint: disable=R0902
             self.tree.vcs.lock(self.path)
         # Edit the whole file in an editor
         if edit_all:
+            self.save()
             editor.edit(self.path, tool=tool)
+            self.load(True)
         # Edit only the text part in an editor
         else:
             # Edit the text in a temporary file
@@ -522,10 +584,6 @@ class Item(BaseValidatable, BaseFileObject):  # pylint: disable=R0902
             )
             # Save the text in the actual item file
             self.text = edited_text
-            self.save()
-
-        # Force reloaded
-        self._loaded = False
 
     @auto_save
     def link(self, value):
@@ -551,184 +609,8 @@ class Item(BaseValidatable, BaseFileObject):  # pylint: disable=R0902
         except KeyError:
             log.warning("link to {0} does not exist".format(uid))
 
-    def get_issues(
-        self, skip=None, document_hook=None, item_hook=None
-    ):  # pylint: disable=unused-argument
-        """Yield all the item's issues.
-
-        :param skip: list of document prefixes to skip
-
-        :return: generator of :class:`~doorstop.common.DoorstopError`,
-                              :class:`~doorstop.common.DoorstopWarning`,
-                              :class:`~doorstop.common.DoorstopInfo`
-
-        """
-        assert document_hook is None
-        assert item_hook is None
-        skip = [] if skip is None else skip
-
-        log.info("checking item %s...", self)
-
-        # Verify the file can be parsed
-        self.load()
-
-        # Skip inactive items
-        if not self.active:
-            log.info("skipped inactive item: %s", self)
-            return
-
-        # Delay item save if reformatting
-        if settings.REFORMAT:
-            self.auto = False
-
-        # Check text
-        if not self.text:
-            yield DoorstopWarning("no text")
-
-        # Check external references
-        if settings.CHECK_REF:
-            try:
-                self.find_ref()
-            except DoorstopError as exc:
-                yield exc
-
-        # Check links
-        if not self.normative and self.links:
-            yield DoorstopWarning("non-normative, but has links")
-
-        # Check links against the document
-        yield from self._get_issues_document(self.document, skip)
-
-        if self.tree:
-            # Check links against the tree
-            yield from self._get_issues_tree(self.tree)
-
-            # Check links against both document and tree
-            yield from self._get_issues_both(self.document, self.tree, skip)
-
-        # Check review status
-        if not self.reviewed:
-            if settings.CHECK_REVIEW_STATUS:
-                if not self._data['reviewed']:
-                    if settings.REVIEW_NEW_ITEMS:
-                        self.review()
-                    else:
-                        yield DoorstopInfo("needs initial review")
-                else:
-                    yield DoorstopWarning("unreviewed changes")
-
-        # Reformat the file
-        if settings.REFORMAT:
-            log.debug("reformatting item %s...", self)
-            self.save()
-
-    def _get_issues_document(self, document, skip):
-        """Yield all the item's issues against its document."""
-        log.debug("getting issues against document...")
-
-        if document in skip:
-            log.debug("skipping issues against document %s...", document)
-            return
-
-        # Verify an item's UID matches its document's prefix
-        if self.prefix != document.prefix:
-            msg = "prefix differs from document ({})".format(document.prefix)
-            yield DoorstopInfo(msg)
-
-        # Verify that normative, non-derived items in a child document have at
-        # least one link.  It is recommended that these items have an upward
-        # link to an item in the parent document, however, this is not
-        # enforced.  An info message is generated if this is not the case.
-        if all((document.parent, self.normative, not self.derived)) and not self.links:
-            msg = "no links to parent document: {}".format(document.parent)
-            yield DoorstopWarning(msg)
-
-        # Verify an item's links are to the correct parent
-        for uid in self.links:
-            try:
-                prefix = uid.prefix
-            except DoorstopError:
-                msg = "invalid UID in links: {}".format(uid)
-                yield DoorstopError(msg)
-            else:
-                if document.parent and prefix != document.parent:
-                    # this is only 'info' because a document is allowed
-                    # to contain items with a different prefix, but
-                    # Doorstop will not create items like this
-                    msg = "parent is '{}', but linked to: {}".format(
-                        document.parent, uid
-                    )
-                    yield DoorstopInfo(msg)
-
-    def _get_issues_tree(self, tree):
-        """Yield all the item's issues against its tree."""
-        log.debug("getting issues against tree...")
-
-        # Verify an item's links are valid
-        identifiers = set()
-        for uid in self.links:
-            try:
-                item = tree.find_item(uid)
-            except DoorstopError:
-                identifiers.add(uid)  # keep the invalid UID
-                msg = "linked to unknown item: {}".format(uid)
-                yield DoorstopError(msg)
-            else:
-                # check the linked item
-                if not item.active:
-                    msg = "linked to inactive item: {}".format(item)
-                    yield DoorstopInfo(msg)
-                if not item.normative:
-                    msg = "linked to non-normative item: {}".format(item)
-                    yield DoorstopWarning(msg)
-                # check the link status
-                if uid.stamp == Stamp(True):
-                    uid.stamp = item.stamp()
-                elif not str(uid.stamp) and settings.STAMP_NEW_LINKS:
-                    uid.stamp = item.stamp()
-                elif uid.stamp != item.stamp():
-                    if settings.CHECK_SUSPECT_LINKS:
-                        msg = "suspect link: {}".format(item)
-                        yield DoorstopWarning(msg)
-                # reformat the item's UID
-                identifier2 = UID(item.uid, stamp=uid.stamp)
-                identifiers.add(identifier2)
-
-        # Apply the reformatted item UIDs
-        if settings.REFORMAT:
-            self._data['links'] = identifiers
-
-    def _get_issues_both(self, document, tree, skip):
-        """Yield all the item's issues against its document and tree."""
-        log.debug("getting issues against document and tree...")
-
-        if document.prefix in skip:
-            log.debug("skipping issues against document %s...", document)
-            return
-
-        # Verify an item is being linked to (child links)
-        if settings.CHECK_CHILD_LINKS and self.normative:
-            find_all = settings.CHECK_CHILD_LINKS_STRICT or False
-            items, documents = self._find_child_objects(
-                document=document, tree=tree, find_all=find_all
-            )
-
-            if not items:
-                for child_document in documents:
-                    if document.prefix in skip:
-                        msg = "skipping issues against document %s..."
-                        log.debug(msg, child_document)
-                        continue
-                    msg = "no links from child document: {}".format(child_document)
-                    yield DoorstopWarning(msg)
-            elif settings.CHECK_CHILD_LINKS_STRICT:
-                prefix = [item.prefix for item in items]
-                for child in document.children:
-                    if child in skip:
-                        continue
-                    if child not in prefix:
-                        msg = 'no links from document: {}'.format(child)
-                        yield DoorstopWarning(msg)
+    def is_reviewed(self):
+        return self._data['reviewed']
 
 
 
@@ -753,37 +635,40 @@ class Item(BaseValidatable, BaseFileObject):  # pylint: disable=R0902
         if not settings.CACHE_PATHS:
             cached_files = {}
         # Search for the external reference
-        log.debug("seraching for ref '{}'...".format(self.ref))
-        pattern = r"(\b|\W){}(\b|\W)".format(re.escape(self.ref))
-        log.trace("regex: {}".format(pattern))  # type: ignore
-        regex = re.compile(pattern)
-        for path, filename, relpath in self.tree.vcs.paths:
-            # Skip the item's file while searching
-            if path == self.path:
-                continue
-            # Check for a matching filename
-            if filename == self.ref:
-                return relpath, None
-            # Skip extensions that should not be considered text
-            if os.path.splitext(filename)[-1] in settings.SKIP_EXTS:
-                continue
-            # Search for the reference in the file
-            lines = self.tree.getlines(path)
-            if lines is None:
-                log.trace("unable to read lines from: {}".format(path))  # type: ignore
-                continue
-            lineno = 1
-            for line in lines:
-                try:
-                    if regex.search(line.decode('utf-8')):
-                        log.debug("found ref: {}".format(relpath))
-                        return relpath, lineno
-                except:
-                    pass
-                lineno += 1
+        return self.reference_finder.find_ref(self.ref, self.tree, self.path)
 
-        msg = "external reference not found: {}".format(self.ref)
-        raise DoorstopError(msg)
+    @requires_tree
+    def find_references(self):
+        """Get the array of references. Check each references before returning.
+
+        :raises: :class:`~doorstop.common.DoorstopError` when no
+            reference is found
+
+        :return: Array of tuples:
+            (
+              relative path to file or None (when no reference set),
+              line number (when found in file) or None (when found as
+              filename) or None (when no reference set)
+            )
+
+        """
+
+        if not self.references:
+            log.debug("no external reference to search for")
+            return []
+        if not settings.CACHE_PATHS:
+            pyficache.clear_file_cache()
+
+        references = []
+        for ref_item in self.references:
+            path = ref_item["path"]
+            keyword = ref_item["keyword"] if "keyword" in ref_item else None
+
+            reference = self.reference_finder.find_file_reference(
+                path, self.root, self.tree, path, keyword
+            )
+            references.append(reference)
+        return references
 
     def find_child_links(self, find_all=True):
         """Get a list of item UIDs that link to this item (reverse links).
@@ -793,7 +678,7 @@ class Item(BaseValidatable, BaseFileObject):  # pylint: disable=R0902
         :return: list of found item UIDs
 
         """
-        items, _ = self._find_child_objects(find_all=find_all)
+        items, _ = self.find_child_items_and_documents(find_all=find_all)
         identifiers = [item.uid for item in items]
         return identifiers
 
@@ -807,7 +692,7 @@ class Item(BaseValidatable, BaseFileObject):  # pylint: disable=R0902
         :return: list of found items
 
         """
-        items, _ = self._find_child_objects(find_all=find_all)
+        items, _ = self.find_child_items_and_documents(find_all=find_all)
         return items
 
     child_items = property(find_child_items)
@@ -818,12 +703,12 @@ class Item(BaseValidatable, BaseFileObject):  # pylint: disable=R0902
         :return: list of found documents
 
         """
-        _, documents = self._find_child_objects(find_all=False)
+        _, documents = self.find_child_items_and_documents(find_all=False)
         return documents
 
     child_documents = property(find_child_documents)
 
-    def _find_child_objects(self, document=None, tree=None, find_all=True):
+    def find_child_items_and_documents(self, document=None, tree=None, find_all=True):
         """Get lists of child items and child documents.
 
         :param document: document containing the current item
@@ -872,15 +757,15 @@ class Item(BaseValidatable, BaseFileObject):  # pylint: disable=R0902
     def stamp(self, links=False):
         """Hash the item's key content for later comparison."""
         values = [self.uid, self.text, self.ref]
+
+        if self.references:
+            values.append(self.references)
+
         if links:
             values.extend(self.links)
         for key in self.document.extended_reviewed:
             if key in self._data:
-                values.append(self._dump(self._data[key]))
-            else:
-                log.warning(
-                    "{}: missing extended reviewed attribute: {}".format(self.uid, key)
-                )
+                values.append(_convert_to_str(self._data[key], ""))
         return Stamp(*values)
 
     @auto_save
@@ -931,9 +816,6 @@ class UnknownItem:
     def uid(self):
         """Get the item's UID."""
         return self._uid
-
-    prefix = Item.prefix
-    number = Item.number
 
     @property
     def relpath(self):